tests/wpt/tests/css/css-text-decor/parsing/text-decoration-color-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Text Decoration Test: Parsing text-decoration-color with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-3/#text-decoration-color-property">
<meta name="assert" content="text-decoration-color supports only the grammar '<color>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("text-decoration-color", "auto");
test_invalid_value("text-decoration-color", "invert");
test_invalid_value("text-decoration-color", "50%");
test_invalid_value("text-decoration-color", "red green");
</script>
